fix: traced_getfield move to TracedUtils
