+++ src/Compiler.jl
@@ -13,11 +13,7 @@ import ..Reactant:
     TracedToConcrete,
     append_path,
     TracedType
-
-@inline traced_getfield(@nospecialize(obj), field) = Base.getfield(obj, field)
-@inline traced_getfield(
-    @nospecialize(obj::AbstractArray{<:Union{ConcreteRNumber,ConcreteRArray}}), field
-) = Base.getindex(obj, field)
+import ..TracedUtils: TracedUtils, traced_getfield
 
 function create_result(tocopy::T, path, result_stores) where {T}
     if !isstructtype(typeof(tocopy))
