+++ src/TracedUtils.jl
@@ -14,11 +14,17 @@ using ..Reactant:
     AnyTracedRArray,
     MissingTracedValue,
     OrderedIdDict,
-    Compiler
+    ConcreteRArray,
+    ConcreteRNumber
 import ..Reactant
 import ..Reactant.MLIR
 import ..ReactantPrimitive
 import ..Ops
+
+@inline traced_getfield(@nospecialize(obj), field) = Base.getfield(obj, field)
+@inline traced_getfield(
+    @nospecialize(obj::AbstractArray{<:Union{ConcreteRNumber,ConcreteRArray}}), field
+) = Base.getindex(obj, field)
 
 materialize_traced_array(x::TracedRArray) = x
 materialize_traced_array(x::WrappedTracedRArray) = x[axes(x)...]
